add new_hdfs_file_reader 2
