+++ be/src/io/CMakeLists.txt
@@ -26,7 +26,6 @@ set(IO_FILES
     broker_writer.cpp
     buffered_reader.cpp
     file_factory.cpp
-    new_file_factory.cpp
     hdfs_builder.cpp
     hdfs_file_reader.cpp
     hdfs_reader_writer.cpp
